Refactor widgets/header/theme_switcher
