+++ src/shared/lib/hooks/useCounter.ts
