+++ src/shared/lib/hooks/useInterval.ts
